Add `cmake` variable with default cmake flags
Before this commit, cmake was called directly from orchestra components.
This fact made difficult to provide a set of flags for cmake to be used
by default by all cmake projects built with orchestra.

For our builds there are a set of cmake flags that we always want to
pass. Currently these flags are two (CMAKE_INSTALL_PREFIX and
CMAKE_EXPORT_COMPILE_COMMANDS), but we might want to add more in the
future.

This commit defines a `cmake_default_cmdline_extra_flags` function and a
`cmake` variable.
The `cmake_default_cmdline_extra_flags` is the function that wraps all
the cmake default flags that we want to pass to all cmake projects.
The `cmake` variable is a wrapper that allows orchestra components to
call cmake with all the default flags specified in
`cmake_default_cmdline_extra_flags`.

This commit updates all the cmake-based components, so that they now call
(@= cmake @) instead of cmake, getting the new default arguments as
expected.

This also fixes an issue with cmake-based components not generating
compile_commands.json files as we expected.