Add `cmake` variable with default CMake flags
Before this commit, cmake was called directly from orchestra components.
This fact made difficult to provide a set of flags for cmake to be used
by default by all cmake projects built with orchestra.

For our builds there are a set of cmake flags that we always want to
pass. Currently these flags are three (DCMAKE_INSTALL_SO_NO_EXE,
CMAKE_INSTALL_PREFIX and CMAKE_EXPORT_COMPILE_COMMANDS), but we might
want to add more in the future.

This commit updates all the cmake-based components, so that they now call
(@= cmake @) instead of cmake, getting the new default arguments as
expected.

This also fixes an issue with cmake-based components not generating
compile_commands.json files as we expected.