Make -rpath a "late" argument
The introduction of HARD_FLAGS_* led us having additional arguments
specified immediately at the beginning of the command line. In turn,
this leads to having an `-rpath` flag specified there having top
priority. This is not OK since it overrides (in terms of priority)
relative rpaths in build directories, ultimately leading binaries to use
installed libraries instead of those available in the build directory.

This commit introduces the possibility for
`hard-flags-compiler-wrapper.py` to specify "late" options, which are
appended to the command line instead of being put at the beginning (the
default).