Create target branch if it is not default
If the branch that was pushed to trigger the CI was not part of the
default list (develop, master and the related next-* counterparts) then
that branch is created in orchestra as well. This is needed to create
the correct binary archives symlinks.

Symlinks have the following structure:
<orchestra_branch_name>_<component_branch_name>.tar.gz -> binarch.tar.gz

Where orchestra_branch_name is the name of the checked out branch for
orchestra at the time the symlink is created. The same goes for
component_branch_name.

To understand why this change is needed, consider this: pushing to
revng:somebranch also rebuilds binary archives for all revng
dependencies and update their symlinks. Assuming orchestra config does
not have a branch named "somebranch", the default branch will get
checked out (master), leading to updating the symlinks named
master_<component_branch_name>.tar.gz.
This means that, for example, the symlink for revng-c
master_master.tar.gz would end up pointing to a binary archive for
revng-c built with the "somebranch" version of revng.

This is a problem because:

1. the symlink points to an archive which is not "valid", because it
   should be pointing to the binary archive for revng-c built using the
   default version of its dependencies
2. the symlink will point to a newly built archive and the previously
   pointed to archive might not be referenced by any symlink anymore.
   Unreferenced are deleted, so orchestra users will not find archives
   for revng dependencies