llvm-documentation: enable figures in SVG
SVGs figures are supported by default in `graphviz`, and they are
strictly better than png, because they have searchable text in them.

Moreover, this fixes a subtle bug: if you try to build
llvm-documentation in an orchestra environment where the `ui/graphviz`
is installed, building the llvm-ducumentation will try to used that and
fail because we ship `graphviz` and configuring it with
`--without-libgd`, so the `dot` utility we install cannot emit png files.

SVG is builting in `graphviz`, so switching to SVG enables to
successfully build llvm-documentation even on systems where
`ui/graphviz` is installed.