Enable javascript in llvm-documentation Info.plist
This allows opening the drop-down visualization of figures in Zeal,
which otherwise seems to be disabled by default in some versions.

For now, this is done with a violent invocation of `sed` on the
`Info.plist` file itself, because that file is generated directly by
Doxygen which currently does not provided options for its customization.

There could be better solution but for now they're not worth the effort.