Javascript components: check re-configure
Before this commit the Javascript-based components (vscode-web,
revng-ui) would download their packages during the `configure` phase and
never check during install if the lockfile had changed. This commit
introduces this check where if the package lock of the component changes
it will cause the `install` phase to be interrupted with an informative
message.