ci: skip creating feature branch
According to commit 960e5ae the creation of a branch for `feature/*`
pushes was needed to allow creating the correct symlinks in binary
archives. Since now the binary archives symlinks are created explicitly
via `orc symlink-binary-archives`, this is no longer needed.
This also fixes a bug where the `ogit checkout ... master` command would
fail because the `master` branch is not the default and so it's not
checked out from the remote.